+++ libs/Texturize.Sampling/src/SynthesisSettings.cpp
@@ -8,12 +8,12 @@ using namespace Texturize;
 ///// Synthesizer Settings implementation                                                     /////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
-SynthesisSettings::SynthesisSettings(uint64_t rngState) :
+SynthesisSettings::SynthesisSettings(unsigned int rngState) :
 	SynthesisSettings(cv::Point2f(0.f, 0.f))
 {
 }
 
-SynthesisSettings::SynthesisSettings(cv::Point2f seedCoords, int kernel, uint64_t rngState) :
+SynthesisSettings::SynthesisSettings(cv::Point2f seedCoords, int kernel, unsigned int rngState) :
 	_seedKernel(kernel), _seedCoords(seedCoords), _rngState(rngState)
 {
 }
@@ -30,10 +30,10 @@ bool SynthesisSettings::validate() const
 	return valid;
 }
 
-SynthesisSettings SynthesisSettings::random(int kernel, uint64_t state)
+SynthesisSettings SynthesisSettings::random(int kernel, unsigned int state)
 {
 	// Randomly generate the seed coords.
-	cv::RNG rng = cv::RNG(state);
+	cv::RNG rng = cv::RNG(static_cast<unsigned long int>(state));
 
 	return SynthesisSettings(cv::Point2f(rng.uniform(0.f, 1.f), rng.uniform(0.f, 1.f)), kernel, state);
 }
@@ -42,37 +42,37 @@ SynthesisSettings SynthesisSettings::random(int kernel, uint64_t state)
 ///// Pyramid Synthesizer Settings implementation                                             /////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
-PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, float randomness, uint64_t rngState) :
+PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, float randomness, unsigned int rngState) :
 	PyramidSynthesisSettings(scale, cv::Point2f(0.f, 0.f), randomness, rngState)
 {
 }
 
-PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, const std::vector<float>& randomness, uint64_t rngState) :
+PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, const std::vector<float>& randomness, unsigned int rngState) :
 	PyramidSynthesisSettings(scale, cv::Point2f(0.f, 0.f), randomness, rngState)
 {
 }
 
-PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, RandomnessSelectorFunction fn, uint64_t rngState) :
+PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, RandomnessSelectorFunction fn, unsigned int rngState) :
 	PyramidSynthesisSettings(scale, cv::Point2f(0.f, 0.f), fn, rngState)
 {
 }
 
-PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, cv::Point2f seedCoords, float randomness, int kernel, uint64_t rngState) :
+PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, cv::Point2f seedCoords, float randomness, int kernel, unsigned int rngState) :
 	PyramidSynthesisSettings(scale, cv::Point2f(0.f, 0.f), [randomness](int, const cv::Mat&) { return randomness; }, rngState)
 {
 }
 
-PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, cv::Point2f seedCoords, const std::vector<float>& randomness, int kernel, uint64_t rngState) :
+PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, cv::Point2f seedCoords, const std::vector<float>& randomness, int kernel, unsigned int rngState) :
 	PyramidSynthesisSettings(scale, cv::Point2f(0.f, 0.f), [randomness](int l, const cv::Mat&) { return l >= randomness.size() ? 0.5f : randomness[l]; }, rngState)
 {
 }
 
-PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, cv::Point2f seedCoords, RandomnessSelectorFunction fn, int kernel, uint64_t rngState) :
+PyramidSynthesisSettings::PyramidSynthesisSettings(float scale, cv::Point2f seedCoords, RandomnessSelectorFunction fn, int kernel, unsigned int rngState) :
 	SynthesisSettings(seedCoords, kernel, rngState), _scale(scale), _randomnessSelector(fn)
 {
 }
 
-PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, float randomness, int kernel, uint64_t state)
+PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, float randomness, int kernel, unsigned int state)
 {
 	// Randomly generate the seed coords.
 	cv::RNG rng = cv::RNG(state);
@@ -80,7 +80,7 @@ PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, float ran
 	return PyramidSynthesisSettings(scale, cv::Point2f(rng.uniform(0.f, 1.f), rng.uniform(0.f, 1.f)), randomness, kernel, state);
 }
 
-PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, const std::vector<float>& randomness, int kernel, uint64_t state)
+PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, const std::vector<float>& randomness, int kernel, unsigned int state)
 {
 	// Randomly generate the seed coords.
 	cv::RNG rng = cv::RNG(state);
@@ -88,7 +88,7 @@ PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, const std
 	return PyramidSynthesisSettings(scale, cv::Point2f(rng.uniform(0.f, 1.f), rng.uniform(0.f, 1.f)), randomness, kernel, state);
 }
 
-PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, RandomnessSelectorFunction fn, int kernel, uint64_t state)
+PyramidSynthesisSettings PyramidSynthesisSettings::random(float scale, RandomnessSelectorFunction fn, int kernel, unsigned int state)
 {
 	// Randomly generate the seed coords.
 	cv::RNG rng = cv::RNG(state);
